Add teeui jni wrapper.
This patch adds a thin JNI wrapper around a teeui library which may be
used for rendering into a generic frame buffer. It implements and
forwards two function, one for setting some device parameters, such as
the dimensions of the frame buffer and some conversion factors, and one
for passing the frame buffer and performing the actual rendering.

Test: N/A
Bug: 111446692
Bug: 111451575
Change-Id: Idd87600c7a6dbac550f276e3cc6917b2520090ef