Store latest settings of TeeuiFramebufferizer in user config file
The latest runtime configuration for TeeuiFramebufferizer would be stored in a hidden config file $HOME/.teeui/.config.json, so as indicated by the file name, this config file is in json format. It would be written every time the run time configuration changes, and loaded whenever TeeuiFramebufferizer gets started also created .lock file in the same directory to indicate the tool is running.

Bug: 150469020
Test: make -j 55 TeeuiFramebufferizer && LD_LIBRARY_PATH=./out/host/linux-x86/lib64 TeeuiFramebufferizer

Change-Id: I761b4de743b64b2ca2377a8416d2d90214a59df5